refactor: Update webpack alias in next.config
This commit updates the webpack alias in the next.config file of the frontend app. The alias '@repo' is now resolved to the path '../packages'. This change allows for easier importing of modules from the packages directory.

Also, update tsconfig paths

This commit updates the paths in the tsconfig.json file of the frontend app. The alias '@repo/*' is now resolved to the path '../../packages/*'. This change ensures that the correct paths are used when importing modules from the packages directory.

Update tailwind.config.js

This commit updates the tailwind.config.js file in the tailwindcss-config package. The file now includes an exclusion for the node_modules directory, preventing it from being processed by Tailwind CSS.